moved CampaignLcdvCodeJoin.java into entity/campaign
