+++ src/main/java/com/example/kwbruunauktion/auktionSystem/repository/campaign/CampaignLcdvCodeJoinRepository.java
@@ -1,6 +1,6 @@
 package com.example.kwbruunauktion.auktionSystem.repository.campaign;
 
-import com.example.kwbruunauktion.auktionSystem.entity.CampaignLcdvCodeJoin;
+import com.example.kwbruunauktion.auktionSystem.entity.campaign.CampaignLcdvCodeJoin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
